refactor: reorg DialManager.lua to Dial/
