+++ TTS_xwing/src/Dial/DialManager.lua
